Убраны замечания GCC.

+++ command_parser.cpp
@@ -11,11 +11,11 @@ class CommandParser : public QObject
 private:
 	bool testCommand(const char* command);
 
-	void handler_logDir(CommandType, int paramCount, const char* params);
-	void handler_logDump(CommandType, int paramCount, const char* params);
-	void handler_logPage(CommandType, int paramCount, const char* params);
-	void handler_get(CommandType, int paramCount, const char* params);
-	void handler_set(CommandType, int paramCount, const char* params);
+	void handler_logDir(CommandType, size_t paramCount, const char* params);
+	void handler_logDump(CommandType, size_t paramCount, const char* params);
+	void handler_logPage(CommandType, size_t paramCount, const char* params);
+	void handler_get(CommandType, size_t paramCount, const char* params);
+	void handler_set(CommandType, size_t paramCount, const char* params);
 private slots:
 	void initTestCase();
 	void test();
@@ -33,25 +33,25 @@ bool CommandParser::testCommand(const char* command)
 	return result == Parser::SuccessResult;
 }
 
-void CommandParser::handler_logDir(CommandType, int paramCount, const char*)
+void CommandParser::handler_logDir(CommandType, size_t paramCount, const char*)
 {
 	qDebug() << "log_dir";
 	QVERIFY(paramCount == 0);
 }
 
-void CommandParser::handler_logDump(CommandType, int paramCount, const char*)
+void CommandParser::handler_logDump(CommandType, size_t paramCount, const char*)
 {
 	qDebug() << "log_dump";
 	QVERIFY(paramCount == 0);
 }
 
-void CommandParser::handler_logPage(CommandType, int paramCount, const char*)
+void CommandParser::handler_logPage(CommandType, size_t paramCount, const char*)
 {
 	qDebug() << "log_page";
 	QVERIFY(paramCount == 0);
 }
 
-void CommandParser::handler_get(CommandType, int paramCount, const char* params)
+void CommandParser::handler_get(CommandType, size_t paramCount, const char* params)
 {
 	qDebug() << "get";
 	QVERIFY(paramCount == 1);
@@ -59,7 +59,7 @@ void CommandParser::handler_get(CommandType, int paramCount, const char* params)
 	QVERIFY(QString(Parser::parameter(0, params)) == "param");
 }
 
-void CommandParser::handler_set(CommandType, int paramCount, const char* params)
+void CommandParser::handler_set(CommandType, size_t paramCount, const char* params)
 {
 	qDebug() << "set";
 	QVERIFY(paramCount == 2);
@@ -70,11 +70,11 @@ void CommandParser::handler_set(CommandType, int paramCount, const char* params)
 
 void CommandParser::initTestCase()
 {
-	Parser::registerHandler(LogDirCommand, [this] (CommandType t, int c, const char* p) { handler_logDir(t, c, p); } );
-	Parser::registerHandler(LogDumpCommand, [this] (CommandType t, int c, const char* p) { handler_logDump(t, c, p); } );
-	Parser::registerHandler(LogPageCommand, [this] (CommandType t, int c, const char* p) { handler_logPage(t, c, p); } );
-	Parser::registerHandler(GetCommand, [this] (CommandType t, int c, const char* p) { handler_get(t, c, p); } );
-	Parser::registerHandler(SetCommand, [this] (CommandType t, int c, const char* p) { handler_set(t, c, p); } );
+	Parser::registerHandler(LogDirCommand, [this] (CommandType t, size_t c, const char* p) { handler_logDir(t, c, p); } );
+	Parser::registerHandler(LogDumpCommand, [this] (CommandType t, size_t c, const char* p) { handler_logDump(t, c, p); } );
+	Parser::registerHandler(LogPageCommand, [this] (CommandType t, size_t c, const char* p) { handler_logPage(t, c, p); } );
+	Parser::registerHandler(GetCommand, [this] (CommandType t, size_t c, const char* p) { handler_get(t, c, p); } );
+	Parser::registerHandler(SetCommand, [this] (CommandType t, size_t c, const char* p) { handler_set(t, c, p); } );
 }
 
 void CommandParser::test()
