+++ parser.h
@@ -21,7 +21,7 @@ public:
 	static void registerHandler(CommandType commandType, CommandHandler&& handler);
 	static Result parse(char data, ParserState& state);
 
-	static const char* parameter(int count, const char* parameters);
+	static const char* parameter(size_t count, const char* parameters);
 };
 
 #endif // PARSER_H
